requirements.txt
Pyrebase==3.0.27
bs4==0.0.1
Scrapy==1.3.3
python-slugify==1.2.1